web/requirements/base.txt
Django==1.11.4

# DJANGO
django-debug-toolbar==1.8
django-extensions==1.8.1
djangorestframework==3.6.3
django-mptt==0.8.7
django_imagekit==4.0.1
django-redis==4.8.0
django-digg-paginator==0.1.3
django-choices==1.6.0
django-filter==1.1.0
django-environ==0.4.3
Jinja2==2.9.6
celery==4.1.0
jsonschema==2.6.0
django-hosts==3.0
django-polymorphic==2.0.2
social-auth-app-django==2.1.0
# Временно
django-jsoneditor

# SLUG
Unidecode==0.4.21

requests

pandas
xlrd
xlsxwriter

# ELASTICSEARCH
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
django-rest-elasticsearch==0.3.2

# Raven
raven==6.5.0